Part 2 - Current Progress
This is my current progress for Part 2. Only shows one speed for the red led blinking. When I add breakpoints between each if statement, it seems to quickly change through the speeds until it gives up and goes back to the default speed (im not sure what is happening). I open the registers and there are moments where either all of P2 or TB1 say "Error value cannot be read".